Fix clang-tidy correctness issues from build/tidy.log
Address top-impact findings from scripts/clang-tidy.sh:

- Camera ctor self-assigned origin (parameter shadowed member, member
  was never initialized). Qualify with this->. (cuda-raytracer-9va, P0)
- Implicit-widening multiplications in PBO size math: cast to
  GLsizeiptr/size_t before the multiply to avoid 32-bit overflow on
  large framebuffers. (cuda-raytracer-dfr)
- CLI arg parsing: replace atoi/atof with strtol/strtod helpers that
  validate via errno + endptr and report errors instead of silently
  treating bad input as 0. (cuda-raytracer-kok)
- Delete unused make_triangle/make_rect in Scene.cpp. (cuda-raytracer-hnf)
- Narrowing conversions (cuda-raytracer-jo2): change Window/Quad
  width/height storage from unsigned int to int to match GLsizei,
  eliminating the bulk of GL-boundary narrowings at the source. Add
  targeted casts in Mesh/Rasterizer (size_t->GLsizeiptr),
  MeshBVHBuilder (int->float), ObjLoader (unsigned->int), GLTexture
  (GLenum->GLint), Input (double->float in mouse rotation).
- Modernize/perf bundle (partial, cuda-raytracer-spa): NULL -> nullptr
  in Shader/Quad/Window; std::endl -> "\n" project-wide; tan/sin/cos
  -> tanf/sinf/cosf in Camera.h to avoid float->double->float round-trip.
- Bonus: guard glfwInitHint(GLFW_PLATFORM, ...) on #ifdef GLFW_PLATFORM
  so the project still links against GLFW 3.3 (the macro is 3.4+).

Verified with full GUI build (Linux/CUDA 12.2/GLFW 3.3) and headless
build; CLI parsing smoke-tested for valid + invalid inputs.

+++ src/Input.cpp
@@ -58,9 +58,9 @@ void Input::process_camera_movement(GLFWwindow* window, KernelInfo& kernelInfo, 
 	}
 
 	// pitch
-	rotation.x += y_diff * 30.0f;
+	rotation.x += static_cast<float>(y_diff * 30.0);
 	// yaw
-	rotation.y += x_diff * 30.0f;
+	rotation.y += static_cast<float>(x_diff * 30.0);
 
 	// roll
 	if (isPressed(GLFW_KEY_E)) {
